fix(engine): review feedback + perf optimizations
