+++ packages/lwc-engine/src/framework/restrictions.ts
@@ -1,9 +1,16 @@
 import assert from "./assert";
-import { getPropertyDescriptor, defineProperties, getOwnPropertyNames, forEach, assign, isString, defineProperty } from "./language";
+import { getPropertyDescriptor, defineProperties, getOwnPropertyNames, forEach, assign, isString, defineProperty, isUndefined, ArraySlice } from "./language";
 import { Component } from "./component";
 import { getGlobalHTMLPropertiesInfo, getPropNameFromAttrName } from "./attributes";
 import { isBeingConstructed } from "./invoker";
-import { getShadowRootVM } from "./vm";
+import { getShadowRootVM, getNodeKey, getCustomElementVM, VM, getNodeOwnerKey } from "./vm";
+import {
+    getAttribute,
+    setAttribute,
+    setAttributeNS,
+    removeAttribute,
+    removeAttributeNS,
+} from "./dom-api";
 
 function getNodeRestrictionsDescriptors(node: Node): PropertyDescriptorMap {
     if (process.env.NODE_ENV === 'production') {
@@ -88,13 +95,144 @@ function getShadowRootRestrictionsDescriptors(sr: ShadowRoot): PropertyDescripto
     return descriptors;
 }
 
+// Custom Elements Restrictions:
+// -----------------------------
+
+function getAttributePatched(this: HTMLElement, attrName: string): string | null {
+    if (process.env.NODE_ENV !== 'production') {
+        const vm = getCustomElementVM(this);
+        assertPublicAttributeCollision(vm, attrName);
+    }
+
+    return getAttribute.apply(this, ArraySlice.call(arguments));
+}
+
+function setAttributePatched(this: HTMLElement, attrName: string, newValue: any) {
+    const vm = getCustomElementVM(this);
+    if (process.env.NODE_ENV !== 'production') {
+        assertTemplateMutationViolation(vm, attrName);
+        assertPublicAttributeCollision(vm, attrName);
+    }
+    setAttribute.apply(this, ArraySlice.call(arguments));
+}
+
+function setAttributeNSPatched(this: HTMLElement, attrNameSpace: string, attrName: string, newValue: any) {
+    const vm = getCustomElementVM(this);
+
+    if (process.env.NODE_ENV !== 'production') {
+        assertTemplateMutationViolation(vm, attrName);
+        assertPublicAttributeCollision(vm, attrName);
+    }
+    setAttributeNS.apply(this, ArraySlice.call(arguments));
+}
+
+function removeAttributePatched(this: HTMLElement, attrName: string) {
+    const vm = getCustomElementVM(this);
+    // marking the set is needed for the AOM polyfill
+    if (process.env.NODE_ENV !== 'production') {
+        assertTemplateMutationViolation(vm, attrName);
+        assertPublicAttributeCollision(vm, attrName);
+    }
+    removeAttribute.apply(this, ArraySlice.call(arguments));
+}
+
+function removeAttributeNSPatched(this: HTMLElement, attrNameSpace: string, attrName: string) {
+    const vm = getCustomElementVM(this);
+
+    if (process.env.NODE_ENV !== 'production') {
+        assertTemplateMutationViolation(vm, attrName);
+        assertPublicAttributeCollision(vm, attrName);
+    }
+    removeAttributeNS.apply(this, ArraySlice.call(arguments));
+}
+
+function assertPublicAttributeCollision(vm: VM, attrName: string) {
+    if (process.env.NODE_ENV === 'production') {
+        // this method should never leak to prod
+        throw new ReferenceError();
+    }
+    const propName = isString(attrName) ? getPropNameFromAttrName(attrName.toLocaleLowerCase()) : null;
+    const { def: { props: propsConfig } } = vm;
+
+    if (propsConfig && propName && propsConfig[propName]) {
+        assert.logError(`Invalid attribute "${attrName.toLocaleLowerCase()}" for ${vm}. Instead access the public property with \`element.${propName};\`.`);
+    }
+}
+
+function assertTemplateMutationViolation(vm: VM, attrName: string) {
+    if (process.env.NODE_ENV === 'production') {
+        // this method should never leak to prod
+        throw new ReferenceError();
+    }
+    const { elm } = vm;
+    if (!isAttributeChangeControlled(attrName) && !isUndefined(getNodeOwnerKey(elm))) {
+        assert.logError(`Invalid operation on Element ${vm}. Elements created via a template should not be mutated using DOM APIs. Instead of attempting to update this element directly to change the value of attribute "${attrName}", you can update the state of the component, and let the engine to rehydrate the element accordingly.`);
+    }
+    // attribute change control must be released every time its value is checked
+    resetAttributeChangeControl();
+}
+
+let controlledAttributeChange: boolean = false;
+let controlledAttributeName: string | void;
+
+function isAttributeChangeControlled(attrName: string): boolean {
+    if (process.env.NODE_ENV === 'production') {
+        // this method should never leak to prod
+        throw new ReferenceError();
+    }
+    return controlledAttributeChange && attrName === controlledAttributeName;
+}
+
+function resetAttributeChangeControl() {
+    if (process.env.NODE_ENV === 'production') {
+        // this method should never leak to prod
+        throw new ReferenceError();
+    }
+    controlledAttributeChange = false;
+    controlledAttributeName = undefined;
+}
+
+export function prepareForValidAttributeMutation(elm: Element, key: string) {
+    if (process.env.NODE_ENV === 'production') {
+        // this method should never leak to prod
+        throw new ReferenceError();
+    }
+    if (!isUndefined(getNodeKey(elm))) {
+        // TODO: we should guarantee that the methods of the element are all patched
+        controlledAttributeChange = true;
+        controlledAttributeName = key;
+    }
+}
+
 function getCustomElementRestrictionsDescriptors(elm: HTMLElement): PropertyDescriptorMap {
     if (process.env.NODE_ENV === 'production') {
         // this method should never leak to prod
         throw new ReferenceError();
     }
     const descriptors: PropertyDescriptorMap = getNodeRestrictionsDescriptors(elm);
-    return assign(descriptors, {});
+    return assign(descriptors, {
+        // replacing mutators and accessors on the element itself to catch any mutation
+        getAttribute: {
+            value: getAttributePatched,
+            configurable: true,
+        },
+        setAttribute: {
+            value: setAttributePatched,
+            configurable: true,
+        },
+        setAttributeNS: {
+            value: setAttributeNSPatched,
+            configurable: true,
+        },
+        removeAttribute: {
+            value: removeAttributePatched,
+            configurable: true,
+        },
+        removeAttributeNS: {
+            value: removeAttributeNSPatched,
+            configurable: true,
+        },
+    });
 }
 
 function getComponentRestrictionsDescriptors(cmp: Component): PropertyDescriptorMap {
